[step4] refactor: RaceService -> RaceManager
