+++ src/test/java/carracing/controller/RaceControllerTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import carracing.service.RaceService;
+import carracing.domain.RaceManager;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -15,7 +15,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 public class RaceControllerTest {
 
-    RaceController raceController = new RaceController(new RaceService());
+    RaceController raceController = new RaceController(new RaceManager());
 
     @Test
     @DisplayName("carNames 리스트가 null 인 경우")
